+++ includes/db.php
@@ -47,6 +47,11 @@ function dbQuery($pdo, $statement, $params) {
 			else {
 				$query->execute();
 			}
+      if($query->errorCode() != 0) {
+        $errors = $query->errorInfo();
+        echo " failed: ".$errors[2]."<br>";
+        exit();
+      }
 		}
 		catch (PDOException $e) {
 			echo 'Could not perform query: '.$e->getMessage();