+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added basic installation script to add `report_stats` view from file and attempt to cleanup after itself.
 - Further fleshed out `README.md`
 - Improved compatibility with older SQL databases that do not support `INET6_ATON` or `INET6_NTOA`.
+- Added SQL error output.
 
 ## 0-α8